Cleanup: Removed old sedynemo and renamed directional sedynemo to sedynemo.
