+++ osl_dynamics/models/sedynemo_obs.py
@@ -7,16 +7,16 @@ from dataclasses import dataclass
 import numpy as np
 import tensorflow as tf
 from tensorflow_probability import bijectors as tfb
-from tensorflow.keras import layers
+from tensorflow.keras import layers, initializers
 
 import osl_dynamics.data.tf as dtf
 from osl_dynamics.models import dynemo_obs
 from osl_dynamics.models.mod_base import BaseModelConfig, ModelBase
 from osl_dynamics.inference.layers import (
     LogLikelihoodLossLayer,
+    LearnableTensorLayer,
     VectorsLayer,
     CovarianceMatricesLayer,
-    DiagonalMatricesLayer,
     ConcatEmbeddingsLayer,
     SubjectMapLayer,
     MixSubjectSpecificParametersLayer,
@@ -27,7 +27,9 @@ from osl_dynamics.inference.layers import (
     StaticKLDivergenceLayer,
     KLLossLayer,
     MultiLayerPerceptronLayer,
+    StandardizationLayer,
 )
+import osl_dynamics.inference.initializers as osld_initializers
 
 
 @dataclass
@@ -52,8 +54,6 @@ class Config(BaseModelConfig):
         Initialisation for mean vectors.
     initial_covariances : np.ndarray
         Initialisation for mode covariances.
-    diagonal_covariances : bool
-        Should we learn diagonal mode covariances?
     covariances_epsilon : float
         Error added to mode covariances for numerical stability.
     means_regularizer : tf.keras.regularizers.Regularizer
@@ -68,28 +68,18 @@ class Config(BaseModelConfig):
     mode_embeddings_dim : int
         Number of dimensions for the mode embeddings in the spatial maps encoder.
 
-    dev_inf_n_layers : int
-        Number of layers for the inference MLP for deviations.
-    dev_inf_n_units : int
-        Number of units for the inference MLP for deviations.
-    dev_inf_normalization : str
-        Type of normalization for the inference MLP for deviations.
+    dev_n_layers : int
+        Number of layers for the MLP for deviations.
+    dev_n_units : int
+        Number of units for the MLP for deviations.
+    dev_normalization : str
+        Type of normalization for the MLP for deviations.
         Either None, 'batch' or 'layer'.
-    dev_inf_activation : str
-        Type of activation to use for the inference MLP for deviations.
+    dev_activation : str
+        Type of activation to use for the MLP for deviations.
         E.g. 'relu', 'sigmoid', 'tanh', etc.
-    dev_inf_dropout : float
-        Dropout rate for the inference MLP for deviations.
-    dev_mod_n_layers : int
-        Number of layers for the model MLP for deviations.
-    dev_mod_n_units : int
-        Number of units for the model MLP for deviations.
-    dev_mod_normalization : str
-        Type of normalization for the model MLP for deviations.
-    dev_mod_activation : str
-        Type of activation to use for the model MLP for deviations.
-    dev_mod_dropout : float
-        Dropout rate for the model MLP for deviations.
+    dev_dropout : float
+        Dropout rate for the MLP for deviations.
 
     batch_size : int
         Mini-batch size.
@@ -115,7 +105,6 @@ class Config(BaseModelConfig):
     learn_covariances: bool = None
     initial_means: np.ndarray = None
     initial_covariances: np.ndarray = None
-    diagonal_covariances: bool = False
     covariances_epsilon: float = None
     means_regularizer: tf.keras.regularizers.Regularizer = None
     covariances_regularizer: tf.keras.regularizers.Regularizer = None
@@ -125,16 +114,11 @@ class Config(BaseModelConfig):
     subject_embeddings_dim: int = None
     mode_embeddings_dim: int = None
 
-    dev_inf_n_layers: int = 0
-    dev_inf_n_units: int = None
-    dev_inf_normalization: str = None
-    dev_inf_activation: str = None
-    dev_inf_dropout: float = 0.0
-    dev_mod_n_layers: int = 0
-    dev_mod_n_units: int = None
-    dev_mod_normalization: str = None
-    dev_mod_activation: str = None
-    dev_mod_dropout: float = 0.0
+    dev_n_layers: int = 0
+    dev_n_units: int = None
+    dev_normalization: str = None
+    dev_activation: str = None
+    dev_dropout: float = 0.0
 
     def __post_init__(self):
         self.validate_observation_model_parameters()
@@ -162,19 +146,16 @@ class Config(BaseModelConfig):
                 "n_subjects, subject_embeddings_dim and mode_embeddings_dim must be passed."
             )
 
-        if self.dev_inf_n_layers != 0 and self.dev_inf_n_units is None:
+        if self.dev_n_layers != 0 and self.dev_n_units is None:
             raise ValueError("Please pass dev_inf_n_units.")
-
-        if self.dev_mod_n_layers != 0 and self.dev_mod_n_units is None:
-            raise ValueError("Plase pass dev_mod_n_units.")
 
 
 class Model(ModelBase):
-    """SE-DyNeMo observation model class.
+    """Observation model for directional SE-DyNeMo
 
     Parameters
     ----------
-    config : osl_dynamics.models.dynemo_obs.Config
+    config : osl_dynamics.models.directional_sedynemo_obs.Config
     """
 
     config_type = Config
@@ -183,31 +164,17 @@ class Model(ModelBase):
         """Builds a keras model."""
         self.model = _model_structure(self.config)
 
-    def get_group_means(self):
-        """Get the group means.
-
-        Returns
-        -------
-        group_means : np.ndarray
-            Group means. Shape is (n_modes, n_channels)
-        """
-        return dynemo_obs.get_means(self.model, "group_means")
-
-    def get_group_covariances(self):
-        """Get the group covariances.
-
-        Returns
-        -------
-        group_covariances : np.ndarray
-            Group covariances. Shape is (n_modes, n_channels, n_channels).
-        """
-        return dynemo_obs.get_covariances(self.model, "group_covs")
-
     def get_group_means_covariances(self):
-        """Get the group means and covariances."""
-        group_means = self.get_group_means()
-        group_covariances = self.get_group_covariances()
-        return group_means, group_covariances
+        """Get the group means and covariances of each mode.
+
+        Returns
+        -------
+        means : np.ndarray
+            Mode means for the group. Shape is (n_modes, n_channels).
+        covariances : np.ndarray
+            Mode covariances for the group. Shape is (n_modes, n_channels, n_channels).
+        """
+        return get_group_means_covariances(self.model)
 
     def get_subject_embeddings(self):
         """Get the subject embedding vectors.
@@ -219,45 +186,6 @@ class Model(ModelBase):
         """
         return get_subject_embeddings(self.model)
 
-    def get_mode_embeddings(self):
-        """Get the mode spatial map embeddings.
-
-        Returns
-        -------
-        means_mode_embeddings : np.ndarray
-            Mode embeddings for means. Shape is (n_modes, mode_embeddings_dim).
-        covs_mode_embeddings : np.ndarray
-            Mode embeddings for covs. Shape is (n_modes, mode_embeddings_dim).
-        """
-        return get_mode_embeddings(self.model)
-
-    def get_concatenated_embeddings(self):
-        """Get the concatenated embedding vectors of deviations.
-
-        Returns
-        -------
-        means_embeddings : np.ndarray
-            Embedding vectors for the mean deviations.
-            Shape is (n_subjects, n_modes, subject_embeddings_dim + mode_embeddings_dim).
-        covs_embeddings : np.ndarray
-            Embedding vectors for the covs deviations.
-            Shape is (n_subjects, n_modes, subject_embeddings_dim + mode_embeddings_dim).
-        """
-        return get_concatenated_embeddings(self.model)
-
-    def get_subject_dev(self):
-        """Get the subject specific deviations of means and covs from the group.
-
-        Returns
-        -------
-        means_dev : np.ndarray
-            Deviation of means from the group. Shape is (n_subjects, n_modes, n_channels).
-        covs_dev : np.ndarray
-            Deviation of Cholesky factor of covs from the group.
-            Shaoe is (n_subjects, n_modes, n_channels * (n_channels + 1) // 2).
-        """
-        return get_subject_dev(self.model)
-
     def get_subject_means_covariances(self):
         """Get the means and covariances for each subject
 
@@ -268,7 +196,9 @@ class Model(ModelBase):
         subject_covs : np.ndarray
             Mode covariances for each subject. Shape is (n_subjects, n_modes, n_channels, n_channels).
         """
-        return get_subject_means_covariances(self.model)
+        return get_subject_means_covariances(
+            self.model, self.config.learn_means, self.config.learn_covariances
+        )
 
     def set_regularizers(self, training_dataset):
         """Set the means and covariances regularizer based on the training data.
@@ -292,7 +222,6 @@ class Model(ModelBase):
                 self.model,
                 training_dataset,
                 self.config.covariances_epsilon,
-                self.config.diagonal_covariances,
                 layer_name="group_covs",
             )
 
@@ -331,8 +260,7 @@ class Model(ModelBase):
         dynemo_obs.set_covariances(
             self.model,
             group_covariances,
-            self.config.diagonal_covariances,
-            update_initializer,
+            update_initializer=update_initializer,
             layer_name="group_covs",
         )
 
@@ -343,22 +271,13 @@ def _model_structure(config):
     alpha = layers.Input(shape=(config.sequence_length, config.n_modes), name="alpha")
     subj_id = layers.Input(shape=(config.sequence_length,), name="subj_id")
 
-    # Observation model:
-    # - We use a multivariate normal with a mean vector and covariance matrix for
-    #   each mode as the observation model.
-    # - Each subject has their own mean vectors and covariance matrices for each mode.
-    #   They are near the group means and covariances.
-    # - We calculate the likelihood of generating the training data with alpha
-    #   and the observation model.
-
-    # Definition of layers
-
-    # Subject embedding layer
+    # Subject embedding layers
     subjects_layer = TFRangeLayer(config.n_subjects, name="subjects")
     subject_embeddings_layer = layers.Embedding(
         config.n_subjects, config.subject_embeddings_dim, name="subject_embeddings"
     )
 
+    # Group level observation model parameters
     group_means_layer = VectorsLayer(
         config.n_modes,
         config.n_channels,
@@ -367,90 +286,190 @@ def _model_structure(config):
         config.means_regularizer,
         name="group_means",
     )
-    if config.diagonal_covariances:
-        group_covs_layer = DiagonalMatricesLayer(
-            config.n_modes,
-            config.n_channels,
-            config.learn_covariances,
-            config.initial_covariances,
-            config.covariances_epsilon,
-            config.covariances_regularizer,
-            name="group_covs",
-        )
-    else:
-        group_covs_layer = CovarianceMatricesLayer(
-            config.n_modes,
-            config.n_channels,
-            config.learn_covariances,
-            config.initial_covariances,
-            config.covariances_epsilon,
-            config.covariances_regularizer,
-            name="group_covs",
-        )
-    means_mode_embeddings_layer = layers.Dense(
-        config.mode_embeddings_dim,
-        name="means_mode_embeddings",
-    )
-    covs_mode_embeddings_layer = layers.Dense(
-        config.mode_embeddings_dim,
-        name="covs_mode_embeddings",
-    )
-    means_concat_embeddings_layer = ConcatEmbeddingsLayer(
+    group_covs_layer = CovarianceMatricesLayer(
         config.n_modes,
         config.n_channels,
-        config.n_subjects,
-        name="means_concat_embeddings",
-    )
-    covs_concat_embeddings_layer = ConcatEmbeddingsLayer(
-        config.n_modes,
-        config.n_channels,
-        config.n_subjects,
-        name="covs_concat_embeddings",
+        config.learn_covariances,
+        config.initial_covariances,
+        config.covariances_epsilon,
+        config.covariances_regularizer,
+        name="group_covs",
     )
 
-    # Inference part of the deviation
-    means_dev_inf_input_layer = MultiLayerPerceptronLayer(
-        config.dev_inf_n_layers,
-        config.dev_inf_n_units,
-        config.dev_inf_normalization,
-        config.dev_inf_activation,
-        config.dev_inf_dropout,
-        name="means_dev_inf_input",
-    )
-    means_dev_inf_mu_layer = layers.Dense(config.n_channels, name="means_dev_inf_mu")
-    means_dev_inf_sigma_layer = layers.Dense(
-        config.n_channels, activation="softplus", name="means_dev_inf_sigma"
-    )
+    subjects = subjects_layer(data)
+    subject_embeddings = subject_embeddings_layer(subjects)
+
+    group_mu = group_means_layer(data)
+    group_D = group_covs_layer(data)
+
+    # ---------------
+    # Mean deviations
+
+    # Layer definitions
     if config.learn_means:
-        means_dev_layer = SampleNormalDistributionLayer(
-            config.theta_std_epsilon, name="means_dev"
+        means_mode_embeddings_layer = layers.Dense(
+            config.mode_embeddings_dim,
+            name="means_mode_embeddings",
         )
+        means_concat_embeddings_layer = ConcatEmbeddingsLayer(
+            config.n_modes,
+            config.n_channels,
+            config.n_subjects,
+            name="means_concat_embeddings",
+        )
+
+        means_dev_map_input_layer = MultiLayerPerceptronLayer(
+            config.dev_n_layers,
+            config.dev_n_units,
+            config.dev_normalization,
+            config.dev_activation,
+            config.dev_dropout,
+            name="means_dev_map_input",
+        )
+        means_dev_map_layer = layers.Dense(config.n_channels, name="means_dev_map")
+        norm_means_dev_map_layer = StandardizationLayer(
+            axis=-1, name="norm_means_dev_map"
+        )
+
+        means_dev_mag_mod_sigma_layer = layers.Dense(
+            1, activation="softplus", name="means_dev_mag_mod_sigma"
+        )
+        means_dev_mag_inf_mu_layer = LearnableTensorLayer(
+            shape=(config.n_subjects, config.n_modes, 1),
+            learn=config.learn_means,
+            initializer=initializers.TruncatedNormal(mean=0, stddev=0.02),
+            name="means_dev_mag_inf_mu",
+        )
+        means_dev_mag_inf_sigma_input_layer = LearnableTensorLayer(
+            shape=(config.n_subjects, config.n_modes, 1),
+            learn=config.learn_means,
+            initializer=osld_initializers.SoftplusNormalInitializer(mean=0, std=0.02),
+            name="means_dev_mag_inf_sigma_input",
+        )
+        means_dev_mag_inf_sigma_layer = layers.Activation(
+            "softplus", name="means_dev_mag_inf_sigma"
+        )
+        means_dev_mag_input_layer = SampleNormalDistributionLayer(
+            config.covariances_epsilon, name="means_dev_mag_input"
+        )
+        means_dev_mag_layer = layers.Activation("softplus", name="means_dev_mag")
+
+        means_dev_layer = layers.Multiply(name="means_dev")
+
+        # Data flow to get the subject specific deviations of means
+
+        # Get the concatenated embeddings
+        means_mode_embeddings = means_mode_embeddings_layer(group_mu)
+        means_concat_embeddings = means_concat_embeddings_layer(
+            [subject_embeddings, means_mode_embeddings]
+        )
+
+        # Get the mean deviation maps (no global magnitude information)
+        means_dev_map_input = means_dev_map_input_layer(means_concat_embeddings)
+        means_dev_map = means_dev_map_layer(means_dev_map_input)
+        norm_means_dev_map = norm_means_dev_map_layer(means_dev_map)
+
+        # Get the deviation magnitudes (scale deviation maps globally)
+        means_dev_mag_mod_sigma = means_dev_mag_mod_sigma_layer(means_concat_embeddings)
+        means_dev_mag_inf_mu = means_dev_mag_inf_mu_layer(data)
+        means_dev_mag_inf_sigma_input = means_dev_mag_inf_sigma_input_layer(data)
+        means_dev_mag_inf_sigma = means_dev_mag_inf_sigma_layer(
+            means_dev_mag_inf_sigma_input
+        )
+        means_dev_mag_input = means_dev_mag_input_layer(
+            [means_dev_mag_inf_mu, means_dev_mag_inf_sigma]
+        )
+        means_dev_mag = means_dev_mag_layer(means_dev_mag_input)
+        means_dev = means_dev_layer([means_dev_mag, norm_means_dev_map])
     else:
         means_dev_layer = ZeroLayer(
             shape=(config.n_subjects, config.n_modes, config.n_channels),
             name="means_dev",
         )
+        means_dev = means_dev_layer(data)
 
-    covs_dev_inf_input_layer = MultiLayerPerceptronLayer(
-        config.dev_inf_n_layers,
-        config.dev_inf_n_units,
-        config.dev_inf_normalization,
-        config.dev_inf_activation,
-        config.dev_inf_dropout,
-        name="covs_dev_inf_input",
-    )
-    covs_dev_inf_mu_layer = layers.Dense(
-        config.n_channels * (config.n_channels + 1) // 2, name="covs_dev_inf_mu"
-    )
-    covs_dev_inf_sigma_layer = layers.Dense(
-        config.n_channels * (config.n_channels + 1) // 2,
-        activation="softplus",
-        name="covs_dev_inf_sigma",
-    )
+    # ----------------------
+    # Covariances deviations
+
+    # Layer definitions
     if config.learn_covariances:
-        covs_dev_layer = SampleNormalDistributionLayer(
-            config.theta_std_epsilon, name="covs_dev"
+        covs_mode_embeddings_layer = layers.Dense(
+            config.mode_embeddings_dim,
+            name="covs_mode_embeddings",
         )
+        covs_concat_embeddings_layer = ConcatEmbeddingsLayer(
+            config.n_modes,
+            config.n_channels,
+            config.n_subjects,
+            name="covs_concat_embeddings",
+        )
+
+        covs_dev_map_input_layer = MultiLayerPerceptronLayer(
+            config.dev_n_layers,
+            config.dev_n_units,
+            config.dev_normalization,
+            config.dev_activation,
+            config.dev_dropout,
+            name="covs_dev_map_input",
+        )
+        covs_dev_map_layer = layers.Dense(
+            config.n_channels * (config.n_channels + 1) // 2, name="covs_dev_map"
+        )
+        norm_covs_dev_map_layer = StandardizationLayer(
+            axis=-1, name="norm_covs_dev_map"
+        )
+
+        covs_dev_mag_mod_sigma_layer = layers.Dense(
+            1, activation="softplus", name="covs_dev_mag_mod_sigma"
+        )
+        covs_dev_mag_inf_mu_layer = LearnableTensorLayer(
+            shape=(config.n_subjects, config.n_modes, 1),
+            learn=config.learn_covariances,
+            initializer=initializers.TruncatedNormal(mean=0, stddev=0.02),
+            name="covs_dev_mag_inf_mu_layer",
+        )
+        covs_dev_mag_inf_sigma_input_layer = LearnableTensorLayer(
+            shape=(config.n_subjects, config.n_modes, 1),
+            learn=config.learn_covariances,
+            initializer=osld_initializers.SoftplusNormalInitializer(mean=0, std=0.02),
+            name="covs_dev_mag_inf_sigma_input",
+        )
+        covs_dev_mag_inf_sigma_layer = layers.Activation(
+            "softplus", name="covs_dev_mag_inf_sigma"
+        )
+        covs_dev_mag_input_layer = SampleNormalDistributionLayer(
+            config.covariances_epsilon, name="covs_dev_mag_input"
+        )
+        covs_dev_mag_layer = layers.Activation("softplus", name="covs_dev_mag")
+        covs_dev_layer = layers.Multiply(name="covs_dev")
+
+        # Data flow to get subject specific deviations of covariances
+
+        # Get the concatenated embeddings
+        covs_mode_embeddings = covs_mode_embeddings_layer(
+            InverseCholeskyLayer(config.covariances_epsilon)(group_D)
+        )
+        covs_concat_embeddings = covs_concat_embeddings_layer(
+            [subject_embeddings, covs_mode_embeddings]
+        )
+
+        # Get the covariance deviation maps (no global magnitude information)
+        covs_dev_map_input = covs_dev_map_input_layer(covs_concat_embeddings)
+        covs_dev_map = covs_dev_map_layer(covs_dev_map_input)
+        norm_covs_dev_map = norm_covs_dev_map_layer(covs_dev_map)
+
+        # Get the deviation magnitudes (scale deviation maps globally)
+        covs_dev_mag_mod_sigma = covs_dev_mag_mod_sigma_layer(covs_concat_embeddings)
+        covs_dev_mag_inf_mu = covs_dev_mag_inf_mu_layer(data)
+        covs_dev_mag_inf_sigma_input = covs_dev_mag_inf_sigma_input_layer(data)
+        covs_dev_mag_inf_sigma = covs_dev_mag_inf_sigma_layer(
+            covs_dev_mag_inf_sigma_input
+        )
+        covs_dev_mag_input = covs_dev_mag_input_layer(
+            [covs_dev_mag_inf_mu, covs_dev_mag_inf_sigma]
+        )
+        covs_dev_mag = covs_dev_mag_layer(covs_dev_mag_input)
+        covs_dev = covs_dev_layer([covs_dev_mag, norm_covs_dev_map])
     else:
         covs_dev_layer = ZeroLayer(
             shape=(
@@ -460,112 +479,92 @@ def _model_structure(config):
             ),
             name="covs_dev",
         )
+        covs_dev = covs_dev_layer(data)
 
+    # ----------------------------------------
+    # Add deviations to group level parameters
+
+    # Layer definitions
     subject_means_layer = SubjectMapLayer(
         "means", config.covariances_epsilon, name="subject_means"
     )
     subject_covs_layer = SubjectMapLayer(
         "covariances", config.covariances_epsilon, name="subject_covs"
     )
+
+    # Data flow
+    mu = subject_means_layer([group_mu, means_dev])
+    D = subject_covs_layer([group_D, covs_dev])
+
+    # -----------------------------------
+    # Mix the subject specific paraemters
+    # and get the conditional likelihood
+
+    # Layer definitions
     mix_subject_means_covs_layer = MixSubjectSpecificParametersLayer(
         name="mix_subject_means_covs"
     )
     ll_loss_layer = LogLikelihoodLossLayer(config.covariances_epsilon, name="ll_loss")
 
     # Data flow
-    subjects = subjects_layer(data)  # data not used here
-    subject_embeddings = subject_embeddings_layer(subjects)
-
-    group_mu = group_means_layer(data)  # data not used
-    group_D = group_covs_layer(data)  # data not used
-
-    # spatial map embeddings
-    means_mode_embeddings = means_mode_embeddings_layer(group_mu)
-    covs_mode_embeddings = covs_mode_embeddings_layer(
-        InverseCholeskyLayer(config.covariances_epsilon)(group_D)
-    )
-
-    # Now get the subject specific spatial maps
-    means_concat_embeddings = means_concat_embeddings_layer(
-        [subject_embeddings, means_mode_embeddings]
-    )
-    covs_concat_embeddings = covs_concat_embeddings_layer(
-        [subject_embeddings, covs_mode_embeddings]
-    )
-
-    means_dev_inf_input = means_dev_inf_input_layer(means_concat_embeddings)
-    means_dev_inf_mu = means_dev_inf_mu_layer(means_dev_inf_input)
-    means_dev_inf_sigma = means_dev_inf_sigma_layer(means_dev_inf_input)
-    means_dev = means_dev_layer([means_dev_inf_mu, means_dev_inf_sigma])
-
-    covs_dev_inf_input = covs_dev_inf_input_layer(covs_concat_embeddings)
-    covs_dev_inf_mu = covs_dev_inf_mu_layer(covs_dev_inf_input)
-    covs_dev_inf_sigma = covs_dev_inf_sigma_layer(covs_dev_inf_input)
-    covs_dev = covs_dev_layer([covs_dev_inf_mu, covs_dev_inf_sigma])
-
-    mu = subject_means_layer([group_mu, means_dev])
-    D = subject_covs_layer([group_D, covs_dev])
-
-    # Mix with the mode time course
     m, C = mix_subject_means_covs_layer([alpha, mu, D, subj_id])
     ll_loss = ll_loss_layer([data, m, C])
 
-    # Prior part of deviations
-    means_dev_mod_input_layer = MultiLayerPerceptronLayer(
-        config.dev_inf_n_layers,
-        config.dev_inf_n_units,
-        config.dev_inf_normalization,
-        config.dev_inf_activation,
-        config.dev_inf_dropout,
-        name="means_dev_mod_input",
-    )
-    means_dev_mod_sigma_layer = layers.Dense(
-        config.n_channels, activation="softplus", name="means_dev_mod_sigma"
-    )
-    covs_dev_mod_input_layer = MultiLayerPerceptronLayer(
-        config.dev_inf_n_layers,
-        config.dev_inf_n_units,
-        config.dev_inf_normalization,
-        config.dev_inf_activation,
-        config.dev_inf_dropout,
-        name="covs_dev_mod_input",
-    )
-    covs_dev_mod_sigma_layer = layers.Dense(
-        config.n_channels * (config.n_channels + 1) // 2,
-        activation="softplus",
-        name="covs_dev_mod_sigma",
-    )
+    # ---------
+    # KL losses
 
     if config.learn_means:
-        means_dev_kl_loss_layer = StaticKLDivergenceLayer(
-            config.theta_std_epsilon, name="means_dev_kl_loss"
+        # Layer definitions
+        means_dev_mag_mod_sigma_layer = layers.Dense(
+            1, activation="softplus", name="means_dev_mag_mod_sigma"
+        )
+        means_dev_mag_kl_loss_layer = StaticKLDivergenceLayer(
+            config.covariances_epsilon, name="means_dev_mag_kl_loss"
+        )
+
+        # Data flow
+        means_dev_mag_mod_sigma = means_dev_mag_mod_sigma_layer(means_concat_embeddings)
+        means_dev_mag_kl_loss = means_dev_mag_kl_loss_layer(
+            [
+                data,
+                means_dev_mag_inf_mu,
+                means_dev_mag_inf_sigma,
+                means_dev_mag_mod_sigma,
+            ]
         )
     else:
-        means_dev_kl_loss_layer = ZeroLayer((), name="means_dev_kl_loss")
+        means_dev_mag_kl_loss_layer = ZeroLayer((), name="means_dev_mag_kl_loss")
+        means_dev_mag_kl_loss = means_dev_mag_kl_loss_layer(data)
 
     if config.learn_covariances:
-        covs_dev_kl_loss_layer = StaticKLDivergenceLayer(
-            config.theta_std_epsilon, name="covs_dev_kl_loss"
+        # Layer definitions
+        covs_dev_mag_mod_sigma_layer = layers.Dense(
+            1, activation="softplus", name="covs_dev_mag_mod_sigma"
+        )
+        covs_dev_mag_kl_loss_layer = StaticKLDivergenceLayer(
+            config.covariances_epsilon, name="covs_dev_mag_kl_loss"
+        )
+
+        # Data flow
+        covs_dev_mag_mod_sigma = covs_dev_mag_mod_sigma_layer(covs_concat_embeddings)
+        covs_dev_mag_kl_loss = covs_dev_mag_kl_loss_layer(
+            [
+                data,
+                covs_dev_mag_inf_mu,
+                covs_dev_mag_inf_sigma,
+                covs_dev_mag_mod_sigma,
+            ]
         )
     else:
-        covs_dev_kl_loss_layer = ZeroLayer((), name="covs_dev_kl_loss")
+        covs_dev_mag_kl_loss_layer = ZeroLayer((), name="covs_dev_mag_kl_loss")
+        covs_dev_mag_kl_loss = means_dev_mag_kl_loss_layer(data)
 
-    kl_loss_layer = KLLossLayer(do_annealing=config.do_kl_annealing, name="kl_loss")
+    # Total KL loss
+    # Layer definitions
+    kl_loss_layer = KLLossLayer(config.do_kl_annealing, name="kl_loss")
 
     # Data flow
-    means_dev_mod_input = means_dev_mod_input_layer(means_concat_embeddings)
-    means_dev_mod_sigma = means_dev_mod_sigma_layer(means_dev_mod_input)
-
-    covs_dev_mod_input = covs_dev_mod_input_layer(covs_concat_embeddings)
-    covs_dev_mod_sigma = covs_dev_mod_sigma_layer(covs_dev_mod_input)
-
-    means_dev_kl_loss = means_dev_kl_loss_layer(
-        [data, means_dev_inf_mu, means_dev_inf_sigma, means_dev_mod_sigma]
-    )
-    covs_dev_kl_loss = covs_dev_kl_loss_layer(
-        [data, covs_dev_inf_mu, covs_dev_inf_sigma, covs_dev_mod_sigma]
-    )
-    kl_loss = kl_loss_layer([means_dev_kl_loss, covs_dev_kl_loss])
+    kl_loss = kl_loss_layer([means_dev_mag_kl_loss, covs_dev_mag_kl_loss])
 
     return tf.keras.Model(
         inputs=[data, alpha, subj_id],
@@ -574,65 +573,125 @@ def _model_structure(config):
     )
 
 
+def get_group_means_covariances(model):
+    group_means_layer = model.get_layer("group_means")
+    group_covs_layer = model.get_layer("group_covs")
+
+    group_means = group_means_layer(1)
+    group_covs = group_covs_layer(1)
+    return group_means.numpy(), group_covs.numpy()
+
+
 def get_subject_embeddings(model):
     subject_embeddings_layer = model.get_layer("subject_embeddings")
     n_subjects = subject_embeddings_layer.input_dim
     return subject_embeddings_layer(np.arange(n_subjects)).numpy()
 
 
-def get_mode_embeddings(model):
-    cholesky_bijector = tfb.Chain([tfb.CholeskyOuterProduct(), tfb.FillScaleTriL()])
-
-    group_means = dynemo_obs.get_means(model, "group_means")
-    group_covs = dynemo_obs.get_covariances(model, "group_covs")
-
+def get_means_mode_embeddings(model):
+    group_means, _ = get_group_means_covariances(model)
     means_mode_embeddings_layer = model.get_layer("means_mode_embeddings")
-    covs_mode_embeddings_layer = model.get_layer("covs_mode_embeddings")
-
     means_mode_embeddings = means_mode_embeddings_layer(group_means)
+    return means_mode_embeddings.numpy()
+
+
+def get_covariances_mode_embeddings(model):
+    cholesky_bijector = tfb.Chain([tfb.CholeskyOuterProduct(), tfb.FillScaleTriL()])
+    _, group_covs = get_group_means_covariances(model)
+    covs_mode_embeddings_layer = model.get_layer("covs_mode_embeddings")
     covs_mode_embeddings = covs_mode_embeddings_layer(
         cholesky_bijector.inverse(group_covs)
     )
+    return covs_mode_embeddings.numpy()
 
-    return means_mode_embeddings.numpy(), covs_mode_embeddings.numpy()
 
-
-def get_concatenated_embeddings(model):
+def get_means_concatenated_embeddings(model):
     subject_embeddings = get_subject_embeddings(model)
-    means_mode_embeddings, covs_mode_embeddings = get_mode_embeddings(model)
-
+    means_mode_embeddings = get_means_mode_embeddings(model)
     means_concat_embeddings_layer = model.get_layer("means_concat_embeddings")
-    covs_concat_embeddings_layer = model.get_layer("covs_concat_embeddings")
-
     means_concat_embeddings = means_concat_embeddings_layer(
         [subject_embeddings, means_mode_embeddings]
     )
+    return means_concat_embeddings.numpy()
+
+
+def get_covariances_concatenated_embeddings(model):
+    subject_embeddings = get_subject_embeddings(model)
+    covs_mode_embeddings = get_covariances_mode_embeddings(model)
+    covs_concat_embeddings_layer = model.get_layer("covs_concat_embeddings")
     covs_concat_embeddings = covs_concat_embeddings_layer(
         [subject_embeddings, covs_mode_embeddings]
     )
-    return means_concat_embeddings.numpy(), covs_concat_embeddings.numpy()
+    return covs_concat_embeddings.numpy()
 
 
-def get_subject_dev(model):
-    means_concat_embeddings, covs_concat_embeddings = get_concatenated_embeddings(model)
-    means_dev_inf_input_layer = model.get_layer("means_dev_inf_input")
-    covs_dev_inf_input_layer = model.get_layer("covs_dev_inf_input")
-    means_dev_inf_layer = model.get_layer("means_dev_inf_mu")
-    covs_dev_inf_layer = model.get_layer("covs_dev_inf_mu")
+def get_means_dev_mag(model):
+    means_dev_mag_input_layer = model.get_layer("means_dev_mag_inf_mu")
+    means_dev_mag_layer = model.get_layer("means_dev_mag")
 
-    means_dev_inf_input = means_dev_inf_input_layer(means_concat_embeddings)
-    covs_dev_inf_input = covs_dev_inf_input_layer(covs_concat_embeddings)
-    means_dev = means_dev_inf_layer(means_dev_inf_input)
-    covs_dev = covs_dev_inf_layer(covs_dev_inf_input)
+    means_dev_mag_input = means_dev_mag_input_layer(1)
+    means_dev_mag = means_dev_mag_layer(means_dev_mag_input)
+    return means_dev_mag.numpy()
+
+
+def get_covariances_dev_mag(model):
+    covs_dev_mag_input_layer = model.get_layer("covs_dev_mag_inf_mu")
+    covs_dev_mag_layer = model.get_layer("covs_dev_mag")
+
+    covs_dev_mag_input = covs_dev_mag_input_layer(1)
+    covs_dev_mag = covs_dev_mag_layer(covs_dev_mag_input)
+    return covs_dev_mag.numpy()
+
+
+def get_means_dev_map(model):
+    means_concat_embeddings = get_means_concatenated_embeddings(model)
+
+    means_dev_map_input_layer = model.get_layer("means_dev_map_input")
+    means_dev_map_layer = model.get_layer("means_dev_map")
+    norm_means_dev_map_layer = model.get_layer("norm_means_dev_map")
+
+    means_dev_map_input = means_dev_map_input_layer(means_concat_embeddings)
+    means_dev_map = means_dev_map_layer(means_dev_map_input)
+    norm_means_dev_map = norm_means_dev_map_layer(means_dev_map)
+    return norm_means_dev_map.numpy()
+
+
+def get_covariances_dev_map(model):
+    covs_concat_embeddings = get_covariances_concatenated_embeddings(model)
+
+    covs_dev_map_input_layer = model.get_layer("covs_dev_map_input")
+    covs_dev_map_layer = model.get_layer("covs_dev_map")
+    norm_covs_dev_map_layer = model.get_layer("norm_covs_dev_map")
+
+    covs_dev_map_input = covs_dev_map_input_layer(covs_concat_embeddings)
+    covs_dev_map = covs_dev_map_layer(covs_dev_map_input)
+    norm_covs_dev_map = norm_covs_dev_map_layer(covs_dev_map)
+    return norm_covs_dev_map.numpy()
+
+
+def get_subject_dev(model, learn_means, learn_covariances):
+    means_dev_layer = model.get_layer("means_dev")
+    covs_dev_layer = model.get_layer("covs_dev")
+    if learn_means:
+        means_dev_mag = get_means_dev_mag(model)
+        means_dev_map = get_means_dev_map(model)
+        means_dev = means_dev_layer([means_dev_mag, means_dev_map])
+    else:
+        means_dev = means_dev_layer(1)
+
+    if learn_covariances:
+        covs_dev_mag = get_covariances_dev_mag(model)
+        covs_dev_map = get_covariances_dev_map(model)
+        covs_dev = covs_dev_layer([covs_dev_mag, covs_dev_map])
+    else:
+        covs_dev = covs_dev_layer(1)
 
     return means_dev.numpy(), covs_dev.numpy()
 
 
-def get_subject_means_covariances(model):
-    group_means = dynemo_obs.get_means(model, "group_means")
-    group_covs = dynemo_obs.get_covariances(model, "group_covs")
-
-    means_dev, covs_dev = get_subject_dev(model)
+def get_subject_means_covariances(model, learn_means, learn_covariances):
+    group_means, group_covs = get_group_means_covariances(model)
+    means_dev, covs_dev = get_subject_dev(model, learn_means, learn_covariances)
 
     subject_means_layer = model.get_layer("subject_means")
     subject_covs_layer = model.get_layer("subject_covs")
@@ -644,9 +703,9 @@ def get_subject_means_covariances(model):
 
 def set_bayesian_kl_scaling(model, n_batches, learn_means, learn_covariances):
     if learn_means:
-        means_dev_kl_loss_layer = model.get_layer("means_dev_kl_loss")
-        means_dev_kl_loss_layer.n_batches = n_batches
+        means_dev_mag_kl_loss_layer = model.get_layer("means_dev_mag_kl_loss")
+        means_dev_mag_kl_loss_layer.n_batches = n_batches
 
     if learn_covariances:
-        covs_dev_kl_loss_layer = model.get_layer("covs_dev_kl_loss")
-        covs_dev_kl_loss_layer.n_batches = n_batches
+        covs_dev_mag_kl_loss_layer = model.get_layer("covs_dev_mag_kl_loss")
+        covs_dev_mag_kl_loss_layer.n_batches = n_batches
